FIX : ID 식별값 타입 수정 (Integer to Long)

+++ src/main/java/org/sopt/app/interfaces/postgres/PokeHistoryRepository.java
@@ -3,6 +3,6 @@ package org.sopt.app.interfaces.postgres;
 import org.sopt.app.domain.entity.PokeHistory;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PokeHistoryRepository extends JpaRepository<PokeHistory, Integer> {
+public interface PokeHistoryRepository extends JpaRepository<PokeHistory, Long> {
 
 }
